requirements.txt
py-cord==2.5.0
rasa==3.6.20
python-dotenv==1.0.1